Allow rubinius builds to fail for the time being
A change was introduced into the latest builds of rubinius that exposes
a bug within the multipart-post gem, which incorrectly overrode the
respond_to? method.

A second change was made in rubinius to avoid these problems when
downstream libraries incorrectly override respond_to?. Until that
change makes it into a rubinius release, we'll allow these builds to
fail since the failures aren't because of a problem in hello_sign.

Reference links:
* https://github.com/rubinius/rubinius/commit/9fa7e49ad22d881a6e942a4ec629c119daff6529
* https://github.com/rubinius/rubinius/commit/4c22e32ec8